⚡ Make ProjectDependenciesUpdateForJavaScript a deferred function to avoid loading ncu
